+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 - `Browser`
     - Updated engine to PhantomJS 2.1.1.
     - `#snapshot_id` -- Moved to browser-side `DOMMonitor` for better performance.
+    - `#capture` -- Extract query parameters from `POST` requests.
     - `Javascript`
         - `#dom_elements_with_events`
             - Moved code to browser-side `DOMMonitor`.